+++ settings.gradle
@@ -14,4 +14,5 @@ include(':examples:examples-common')
 include(':examples:example-spring-boot-starter-web')
 include(':examples:example-spring-boot-starter-webflux')
 
+include(':test:openapi-web')
 include(':test:test-utils')
