+++ test/openapi-web/src/main/resources/openapi.yaml
@@ -1,0 +1,1 @@
+../../../../openapi.yaml
